[Update] magnitude, normalization, dot and cross products

+++ src/tuples/tuples.c
@@ -6,13 +6,13 @@
 /*   By: cado-car <cado-car@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/03 23:30:35 by cado-car          #+#    #+#             */
-/*   Updated: 2023/04/04 19:49:05 by cado-car         ###   ########.fr       */
+/*   Updated: 2023/04/05 09:53:23 by cado-car         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_tuple	*tuple(float x, float y, float z, short w)
+t_tuple	*tuple(double x, double y, double z, short w)
 {
 	t_tuple	*tuple;
 
@@ -26,7 +26,7 @@ t_tuple	*tuple(float x, float y, float z, short w)
 	return (tuple);
 }
 
-t_tuple	*point(float x, float y, float z)
+t_tuple	*point(double x, double y, double z)
 {
 	t_tuple	*point;
 
@@ -36,7 +36,7 @@ t_tuple	*point(float x, float y, float z)
 	return (point);
 }
 
-t_tuple	*vector(float x, float y, float z)
+t_tuple	*vector(double x, double y, double z)
 {
 	t_tuple	*vector;
